fix: delete external code and move HOC to another file
